Remove passport and move logic to authController
