+++ server/package.json
@@ -12,10 +12,6 @@
     "express-session": "^1.17.3",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^7.6.1",
-    "passport": "^0.6.0",
-    "passport-jwt": "^4.0.1",
-    "passport-local": "^1.0.0",
-    "passport-local-mongoose": "^8.0.0",
     "validator": "^13.11.0"
   },
   "devDependencies": {
